feat: 修改MosaicInitConfig
